requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
drf-spectacular>=0.27
